requirements.txt
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==2.5.1
mysql-connector-python==8.0.30
python-dotenv==0.20.0
SQLAlchemy==1.4.40
requests==2.28.1